feat(2fa): new InlineTotpSetup using new flow components
Because:

* we need to assemble newly created components into InlineTotpSetup

This commit:

* replaces the existing implementation with a step controller pattern, creates storybooks states and unit tests.
* feature-flags the new design for easy roll back, and updates the functional tests to handle both versions.

Closes #FXA-11720
