+++ packages/fxa-settings/src/pages/InlineTotpSetup/mocks.ts
@@ -2,11 +2,11 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import { MOCK_SESSION_TOKEN, MOCK_UID } from '../mocks';
+import { MOCK_SESSION_TOKEN, MOCK_UID, PLACEHOLDER_QR_CODE } from '../mocks';
 
 export const MOCK_TOTP_TOKEN = {
   secret: 'I5KTS23HJBUDOUTMHBCDCTDZGJSFGTCB',
-  qrCodeUrl: 'data:image/png;base64,pewpewpew',
+  qrCodeUrl: PLACEHOLDER_QR_CODE,
   recoveryCodes: ['recoveryCode1', 'recoveryCode2'],
 };
 export const MOCK_QUERY_PARAMS = {
